+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.4] - 2026-05-08
 ### Added
 - `GHPSError` class
 - `GHPSValidationError` class
@@ -34,7 +35,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Command line interface
 - Test system
 
-[Unreleased]: https://github.com/sepandhaghighi/ghps/compare/v0.3...dev
+[Unreleased]: https://github.com/sepandhaghighi/ghps/compare/v0.4...dev
+[0.4]: https://github.com/sepandhaghighi/ghps/compare/v0.3...v0.4
 [0.3]: https://github.com/sepandhaghighi/ghps/compare/v0.2...v0.3
 [0.2]: https://github.com/sepandhaghighi/ghps/compare/v0.1...v0.2
 [0.1]: https://github.com/sepandhaghighi/ghps/compare/bc0a425...v0.1